TL/UCP: enable linear alltoall (#903)

+++ src/components/tl/ucp/tl_ucp.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (c) 2020-2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  *
  * See file LICENSE for terms.
  */
@@ -60,8 +60,8 @@ typedef struct ucc_tl_ucp_lib_config {
     uint32_t                 scatter_kn_radix;
     ucc_on_off_auto_value_t  scatter_kn_enable_recv_zcopy;
     uint32_t                 scatterv_linear_num_posts;
-    uint32_t                 alltoall_pairwise_num_posts;
-    uint32_t                 alltoallv_pairwise_num_posts;
+    unsigned long            alltoall_pairwise_num_posts;
+    unsigned long            alltoallv_pairwise_num_posts;
     ucc_pipeline_params_t    allreduce_sra_kn_pipeline;
     int                      reduce_avg_pre_op;
     int                      reduce_scatter_ring_bidirectional;
